Day18
Part 2 got me stuck for a while.  I first attempted a true threaded approach, but this was giving 'random' answers (even huge ones above the real result, still not sure how that happened).  I then made a different solution for part 2 that ran them in lockstep, and this gave the right answer (after a little debugging).  I then went back to the threaded version, and debugged it some more, and realized that if a thread was waiting, it didn't get marked as not waiting until it ran (woke up) and set itself as not waiting, but if the first thread ran out of its queue before the other woke up, it would think there was a deadlock.  So I mark the other thread as not waiting when stuff is added to it's queue (alternatively, I could check for waiting and empty queue on other thread when deciding if deadlocked).