Day20
Not too bad.  Had an error in my getDistance() method initially (was getting abs of sum instead of individual components) that slowed me up a bit for part 1.  Took a guess in part 2 that no collisions for 1000 moves should be safe that there would be no more.  Probably could have reduced that to 100, but it worked.