Day 21
This took me a while to get part 1. 
I got the example working after about 50 minutes, but then running my input gave me the wrong answer (which apparently was the right answer for another group of people). 
So I had to debug it, and it turned out my rotate for 2x2 wasn't correct, but because the example only had 1 bit on for the 2x2, the error I had didn't effect it. ( I was doing something like ab/cd -> da/bc ) 
Fixing that issue got me the correct answer for part 1. 

Part 2 was cake. It took me 41 seconds for part 2. And I probably could have cut that nearly in half if my program wasn't printing the state of the image after every iteration! I think it took like 17 seconds to get the answer with the printing happening. After submitting the answer and seeing it was correct, I removed the printing and the solution came in under a second.